chore: rename t6 leaderboardDef folder to leaderboard
